# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Honey School API Gateway
  description: Public contract for the Honey School API Gateway.
  version: 0.1.0
servers:
- url: /api
  description: Honey School web-app API proxy
paths:
...
  /schedule/lessons/{lessonId}:
    get:
      tags:
      - Schedule
      summary: Get scheduled lesson
      description: Returns a scheduled lesson visible to the current user.
      operationId: getScheduledLesson
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Scheduled lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledLessonResponse"
        "401":
          description: Missing or invalid bearer token
        "404":
          description: Scheduled lesson not found
      security:
      - bearerAuth: []
    put:
      tags:
      - Schedule
      summary: Update scheduled lesson
      description: Updates a calendar lesson. Requires TEACHER or ADMIN role.
      operationId: updateScheduledLesson
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledLessonRequest"
        required: true
      responses:
        "200":
          description: Scheduled lesson updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledLessonResponse"
        "400":
          description: Invalid scheduled lesson payload
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Current user cannot manage schedule
        "404":
          description: Scheduled lesson not found
      security:
      - bearerAuth: []
    delete:
      tags:
      - Schedule
      summary: Delete scheduled lesson
      description: Deletes a calendar lesson. Requires TEACHER or ADMIN role.
      operationId: deleteScheduledLesson
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Scheduled lesson deleted
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Current user cannot manage schedule
        "404":
          description: Scheduled lesson not found
      security:
      - bearerAuth: []
...
  /schedule/lessons:
    get:
      tags:
      - Schedule
      summary: List scheduled lessons
      description: "Returns scheduled lessons. Teachers/admins see all, students see\
        \ lessons where they are participants."
      operationId: listScheduledLessons
      responses:
        "200":
          description: Scheduled lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduledLessonResponse"
        "401":
          description: Missing or invalid bearer token
      security:
      - bearerAuth: []
    post:
      tags:
      - Schedule
      summary: Create scheduled lesson
      description: Creates a calendar lesson. Requires TEACHER or ADMIN role.
      operationId: createScheduledLesson
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledLessonRequest"
        required: true
      responses:
        "201":
          description: Scheduled lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledLessonResponse"
        "400":
          description: Invalid scheduled lesson payload
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Current user cannot manage schedule
      security:
      - bearerAuth: []
...
  /schedule/lessons/{lessonId}/start:
    post:
      tags:
      - Schedule
      summary: Start scheduled lesson
      description: Starts a calendar lesson and opens live access. Requires TEACHER
        or ADMIN role.
      operationId: startScheduledLesson
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Scheduled lesson started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledLessonResponse"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Current user cannot manage schedule
        "404":
          description: Scheduled lesson not found
        "409":
          description: Scheduled lesson cannot be started
      security:
      - bearerAuth: []
...
  /schedule/lessons/{lessonId}/participant-links:
    post:
      tags:
      - Schedule
      summary: Get the shared scheduled lesson link
      description: Compatibility endpoint returning the same reusable shared link
        for every participant. Requires TEACHER or ADMIN role.
      operationId: createScheduledLessonParticipantLinks
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Scheduled lesson participant links
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledLessonParticipantLinksResponse"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Current user cannot manage schedule
        "404":
          description: Scheduled lesson not found
      security:
      - bearerAuth: []
...
  /schedule/lessons/{lessonId}/complete:
    post:
      tags:
      - Schedule
      summary: Complete scheduled lesson
      description: Completes a calendar lesson and closes live access. Requires TEACHER
        or ADMIN role.
      operationId: completeScheduledLesson
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Scheduled lesson completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledLessonResponse"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Current user cannot manage schedule
        "404":
          description: Scheduled lesson not found
      security:
      - bearerAuth: []
...
  /schedule/lessons/{lessonId}/schedule:
    patch:
      tags:
      - Schedule
      summary: Reschedule one scheduled lesson
      description: Changes the date and time of only the selected lesson occurrence.
        Requires TEACHER or ADMIN role.
      operationId: rescheduleScheduledLesson
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledLessonScheduleUpdateRequest"
        required: true
      responses:
        "200":
          description: Scheduled lesson rescheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledLessonResponse"
        "400":
          description: Invalid scheduled interval
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Current user cannot manage schedule
        "404":
          description: Scheduled lesson not found
        "409":
          description: Closed lesson cannot be rescheduled
      security:
      - bearerAuth: []
...
  /materials/{materialId}/source-attachments/{attachmentId}/content:
    get:
      tags:
      - material-source-attachment-controller
      summary: Read a teacher-only worksheet source attachment
      operationId: content
      parameters:
      - name: materialId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attachmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
...
  /admin/user-management/operations/{operationId}:
    get:
      tags:
      - user-management-controller
      operationId: operation
      parameters:
      - name: operationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDeletionOperationResponse"
...
components:
  schemas:
...
    ScheduledLessonMaterialAssignmentRequest:
      type: object
      properties:
        materialId:
          type: string
          format: uuid
        participantSubjects:
          type: array
          items:
            type: string
      required:
      - materialId
      - participantSubjects
    ScheduledLessonRecurrenceRequest:
      type: object
      properties:
        mode:
          type: string
          enum:
          - WEEKLY_COUNT
          - WEEKLY_BY_WEEK
        count:
          type: integer
          format: int32
        weekdays:
          type: array
          items:
            type: string
        weekdayTimes:
          type: object
          additionalProperties:
            type: string
        timeZone:
          type: string
      required:
      - count
      - mode
      - timeZone
      - weekdayTimes
      - weekdays
    ScheduledLessonRequest:
      type: object
      properties:
        lessonTemplateId:
          type:
          - string
          - "null"
          format: uuid
        materialId:
          type:
          - string
          - "null"
          format: uuid
        inheritTemplateMaterial:
          type:
          - boolean
          - "null"
        scheduledStart:
          type:
          - string
          - "null"
          format: date-time
        scheduledEnd:
          type:
          - string
          - "null"
          format: date-time
        status:
          type: string
          enum:
          - SCHEDULED
          - IN_PROGRESS
          - COMPLETED
          - CANCELLED
        type:
          type: string
          enum:
          - INDIVIDUAL
          - GROUP
        workMode:
          type: string
          enum:
          - SHARED
          - PARALLEL
        participantSubjects:
          type: array
          items:
            type: string
        participantAssignments:
          type: array
          items:
            $ref: "#/components/schemas/ScheduledLessonMaterialAssignmentRequest"
        recurrence:
          oneOf:
          - $ref: "#/components/schemas/ScheduledLessonRecurrenceRequest"
          - type: "null"
      required:
      - participantAssignments
      - participantSubjects
      - status
      - type
      - workMode
    ScheduledLessonParticipantResponse:
      type: object
      properties:
        subject:
          type: string
        username:
          type:
          - string
          - "null"
        displayName:
          type:
          - string
          - "null"
        attendanceStatus:
          type:
          - string
          - "null"
        materialId:
          type:
          - string
          - "null"
          format: uuid
        materialTitle:
          type:
          - string
          - "null"
      required:
      - subject
    ScheduledLessonResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lessonTemplateId:
          type:
          - string
          - "null"
          format: uuid
        materialId:
          type:
          - string
          - "null"
          format: uuid
        inheritTemplateMaterial:
          type: boolean
        materialTitle:
          type:
          - string
          - "null"
        courseId:
          type:
          - string
          - "null"
          format: uuid
        courseTitle:
          type:
          - string
          - "null"
        lessonTitle:
          type:
          - string
          - "null"
        teacherSubject:
          type:
          - string
          - "null"
        teacherName:
          type:
          - string
          - "null"
        scheduledStart:
          type:
          - string
          - "null"
          format: date-time
        scheduledEnd:
          type:
          - string
          - "null"
          format: date-time
        status:
          type: string
        type:
          type: string
        workMode:
          type: string
        recurrenceSeriesId:
          type:
          - string
          - "null"
          format: uuid
        recurrenceIndex:
          type:
          - integer
          - "null"
          format: int32
        recurrenceTotal:
          type:
          - integer
          - "null"
          format: int32
        livekitRoomName:
          type:
          - string
          - "null"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ScheduledLessonParticipantResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - createdAt
      - id
      - inheritTemplateMaterial
      - participants
      - status
      - type
      - updatedAt
      - workMode
...
    ScheduledLessonParticipantLinkResponse:
      type: object
      properties:
        subject:
          type: string
        displayName:
          type:
          - string
          - "null"
        email:
          type:
          - string
          - "null"
        url:
          type: string
        expiresAt:
          type:
          - string
          - "null"
          format: date-time
        mode:
          type: string
          enum:
          - MAGIC_LINK
          - AUTHENTICATED_LINK
      required:
      - mode
      - subject
      - url
    ScheduledLessonParticipantLinksResponse:
      type: object
      properties:
        lessonId:
          type: string
          format: uuid
        links:
          type: array
          items:
            $ref: "#/components/schemas/ScheduledLessonParticipantLinkResponse"
      required:
      - lessonId
      - links
...
    ScheduledLessonScheduleUpdateRequest:
      type: object
      properties:
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
      required:
      - scheduledEnd
      - scheduledStart
...
    UserDeletionOperationResponse:
      type: object
      properties:
        operationId:
          type: string
          format: uuid
        targetSubject:
          type: string
        status:
          type: string
        errorCode:
          type:
          - string
          - "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type:
          - string
          - "null"
          format: date-time
      required:
      - createdAt
      - operationId
      - status
      - targetSubject
      - updatedAt
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT